Update CreatePosts class

+++ db/migrate/20220712104726_create_posts.rb
@@ -3,8 +3,8 @@ class CreatePosts < ActiveRecord::Migration[7.0]
     create_table :posts do |t|
       t.string :title
       t.string :text
-      t.bigint :comment_counter
-      t.bigint :likes_counter      
+      t.integer :comment_counter
+      t.integer :likes_counter      
       t.references :author, index: false, foreign_key: { to_table: :users }
 
       t.timestamps
